+++ schemacrawler-ai-mcpserver/src/test/java/schemacrawler/tools/ai/mcpserver/server/test/ToolProviderTest.java
@@ -115,14 +115,14 @@ public class ToolProviderTest {
         actualToolNames,
         containsInAnyOrder(
             "about_database",
-            "describe_entities",
-            "describe_relationships",
+            "describe_er_relationships",
             "describe_routines",
             "describe_tables",
             "diagram",
+            "detect_clusters",
             "lint",
             "list",
-            "list_across_tables",
+            "list_members_of_tables",
             "table_sample",
             "table_importance",
             "table_path"));